requirements.txt
# Django
Django==4.0.4
djangorestframework==3.13.1

# Own
requests==2.27.1

# Django related
sqlparse==0.4.2
asgiref==3.5.2
urllib3==1.26.9
charset-normalizer==2.0.12
idna==3.3
certifi==2022.5.18.1
pytz==2020.1